requirements.txt
anyio==4.10.0
APScheduler==3.10.4
certifi==2025.8.3
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
